add equip panel/ add btn logout

+++ views/admin/list-user.html.php
@@ -1,5 +1,11 @@
 <div class="admin-container">
     <h1 class="title">Les clients</h1>
+    <!-- on va afficher les erreurs s'il y en a -->
+    <?php if ($form_result && $form_result->hasErrors()) : ?>
+        <div class="alert alert-danger" role="alert">
+            <?php echo $form_result->getErrors()[0]->getMessage() ?>
+        </div>
+    <?php endif ?>
 
     <table class="table-user">
         <thead>
@@ -19,7 +25,7 @@
                     <td class="footer-description"><?= $user->email ?></td>
                     <td class="footer-description"><?= $user->phone ?></td>
                     <td class="footer-description">
-                        <a class="button-delete" href="/admin/user/delete/<?= $user->id ?>"><i class="bi bi-trash"></i></a>
+                        <a onclick="return confirm('Voulez-vous supprimer cet utilisateur ?')" class="button-delete" href="/admin/user/delete/<?= $user->id ?>"><i class="bi bi-trash"></i></a>
                     </td>
                 </tr>
             <?php endforeach ?>
